backend/timetrackerdice/requirements.txt
django==2.2.1
djangorestframework
markdown
django-filter
TogglPy

